Massive change transitioning to Int12 (and other minor adjustements)
I made an Int12 type for our registers and memory to work with.  It
handles all the arithmatic correctly (I hope) and displays in/reads
from octal form (even though the numbers are signed - is this ok?).

Also, the memory operation "read" has changed to "load" for obvious reasons.

Finally, the "loadProgram" operation has appeared that I hope does the
right thing but is entirely untested and based only on a shakly PDP8
model in my brain (need to verify the incrAddr/incPC - how is our
memory indexed?).